En titty
it only works if the parameters are components are added in the right order. and thats not the easiest thing to set. all the components are packed arrays so theyre always added to the end but theyre iterated over backwards to allow for changes to entities during iteration without messing up the next component.

there is a problem with this method:
1. reparenting cannot currently be done it must be completed in 2 iterations
2. the order that a tree is added is important.

the registry, view, groups and observers are being weird when i try and use them as memebers of a class and only seem to work when theyre static. this is annoying and probably a bug? I hope. my work around is to add a component for each scene or maybe each viewport with a template
if Render<int viewport> is a component I can isolate each viewport using the ECS.

Ive begun working on the scene despite the registry being static i can still isolate the entities that belong to a scene with a SceneTag<> then its up to the user to correctly use the scene tags
I tried making it automatically pick the ID but it has to be compile time constant which is a kinda annoying