repeat after me "I will do my uni work" great. good job. very convincing.
ironed out weird x64 vs x86 glitch which came from differences in how classes are constructed ie win32 does the whole padding thing. I dont understand why. its still given the same size. meta data maybe???

it works now tho which is cool. I need a better way to add new types. which kinda sucks. because theyre template arguments theyre gonna be a prick and require definitions to be in the header file/"translation unit" which basically means when the type is first used the definitions must be in the same file so alternatively I could declare a bunch of types in a seperate file and add that to the .cpp file. the problem with taht is it led to a weird glitch where I would get the error:
"A copy of MyFile.cpp was found in C:\MyDir\MyFile.cpp, but the current source code is different from the version built into C:\MyDir\MyFile.cpp."
while debugging. to fix this you gotta clean and rebuild it everytime you make a change to the .cpp file. I dont understand it but it was really annoying.

I dont understand why there is a difference between classes in x86 and x64. it makes no sense. This is also whu the main loop had to change for the window management.

 Ive also worked out how to share a context. its really not very hard. theres a field for it in the constructor of the window. pretty amazing huh.