the impending demise of the state is upon us worked out how to do shared context and what gets shared. vao among other "state containers" are local to the window. (kinda rly irritating. but oh well.) buffers and "large containers" are shared in the context. yay.
I removed the mess of UniformBuffer i like the inheritance use but overall its just messy and its easier to add it as an associated class rather than an inherited one. to work it into the ECS I'll simply add a member GLbuffer. I guess it could also be inherited but its a bit unneccessary.

I renamed image to texture because im a baller whos ballin'.