'sauce bin 'vamped resource has been revamped. barely. it was entirely for that pun. it now uses a constructor instead of a load function. It kinda makes more sense too. when you use a shared resource you want the pointer but when its unique you might want it to be completely local. there thats an intelligent reason. that'll do.
Appwindow was apparently storing the title twice. not very intelligent. i fixed that.

AppWindow::Init also wasnt resetting itself. so i fixed that. it now takes a collection of windows to display then destroys the context pointer after the window dies. the context pointer is set to null after all the windows are closed so the glfw will be re-initiated.