wiggly lines are crawling up my spines
wierd Transform syntax error where I added an extra speech mark which i have no memory of.

Material holds a map of properties. The properties represents a Texture or a colour and the map indexes property type eg DIFFUSE to the output. The material can is bound in the model.draw() function which now takes a program parameter.

I've added move constructors and deleted copy constructors from texture, model, mesh, shaderProgram and shader. in the hopes to allow the deconstructor to delete the GL objects.

I used a concept and requires(c++ 20) in resource to prevent loading resources without the appropriate type.

I'm having a mental breakdown. -> I had the semicolon moment. but it caused tremendous pain and suffering. there was a missing semicolon in a header file. the header file was included in a cpp file. the next line in the cpp file was include glew. glew needs to be included before glfw or glfw throws a fit. glfw threw 3800+ errors worth of fit. The homocidal rage is contained for now. next time I may not be so strong.

about 25-30 hours of work I thought I was ready for a push. then I was mmm I'll just do this lil bit more...

most of what is written up there ^ is now not true. material holds a map of textures and a uniform buffer. the uniform buffer is indexed with a static class which calculates the correct offsets. arrays and matrices dont work yet becuase theyre optimized differently. the uniform buffer contains a value for each map type and a boolean. not a huge fan of the whole setup but it works so I'll shut up. there are c++ specific things for controlling alignment that I need to look into.

I worked out how to do move constructors to prevent opengl handles being destroyed at the wrong time and added them to all the relevant classes. This means I can use the objects more freely without wrapping it in a shared_ptr.

I also added a GLbuffer class. again. might remove it again.

started using trello again.

found an article about embedding c++ resources into the exe file that Im tempted to try. It looks kinda cool.

I'm going to clean this up a bit and then try and make the render system more componenty.

lets be honest, there are probably more changes.