the slow decay of happy thoughts GLbuffer uses named buffer because thats a thing. means less annoying binding.
brrrrrrrr

completely lost track of the changes made. last looked at this months ago.

scene has been split into the various systems using funky diamond virtual protected inheritance thingies.

material got replaced by specular material. it now uses c++20 wierd concept thingies for the maps. idk if thats rly necessary or a good use but hey what ever. I'm likely gonna replace material again with a super material which will take uniform buffers and allocate them a base to bind to.

The vbos are now components added to each entity and when the entity fulfills an archetype, the vao in the material is initiated. This is done with entt events but could probably be done better by setting up a view or group.

the vbo object needs to be changed somewhat to allow either for copying or for sharing of resource. I can do this by storing a pointer to the first and next shared resource as a form of reference counting.

There is currently no error handling for incomplete mesh structures.

I moved the entt macros to a their own header file.

model loading has been removed. the old code is in notes.txt