we hootin' we hollerin' le owl boi is back mesh loading is up again. I've noticed the assimp mesh loader does little to optimize the mesh produced. The issues as I see them are the scene graph which will get meshes as nodes when really I want them as components of an entity. thats how my program is set up to expect entities. A solution to this could be to distinguish nodes from components with whether it has a transform node. this isnt perfect but idk how to fix the utter garbage I get from assimp.
GLbuffer now destroys itself properly. however I've noticed when the program is closed through the console it doesnt call the destructors. It might be better to fully embrace the resource pattern but idk if its worth it.

Because I'm nearing the bone and animation work I should review the hierarchy component system and make sure it works properly and includes the necessary functionality.

apparently the set parameter functions on texture have been wrong for a really long time. oopsies.