buh brrrr black sheep have you any wool hierarchy-transform system now supports reparenting function. order is maintained by swapping any child with its parent until all the entities are in order. which is deceptively simple and isnt fully tested.
the console now hides itself when in release mode.

load system no longer loads the scene graph or transforms and instead only loads the meshes, materials and textures. This combines the meshes onto and emplaces a singular mesh onto an entity and simply uses the first material. I'm considering using an array to store the materials and then some kind of smart indexing to only render sections of the vertices with the associated material