Cullian Murphy Culling baybee I've begun working on forward+ rendering. so I've added 3 new shaders for the depth pass, the light culling and the light_accumulation.
in shader program: I deleted a dumb thing and added a new constructor for compute shaders

in AppWindow I've changed a bunch of function names to be less ugly. I have a distinct feeling Ive changed them to before. I also removed a thing that I think solved an issue. but i cant recreate the issue so I dont see the problem.

I am seriously reconsidering the AppWindow. I think I'm going to simplify and tidy up alot of the functionality. because its big and ugly.

I am also thinking about a couple of the resource objects. things like "Do I need it" cant I just leave it for the user to manage. which I think yes. Its pretty self contained code tho. I was reading a thing ages ago which said the best way to approach OpenGL objects is very much not to wrap them which I realised I've done. alot. so things like texture have no functionality extending past that of OpenGL. in fact it mostly just limits what you can do.

A replacement for the wrapped class is to simply use a string which. Which accesses the data in the with resource.get("filepath/.png") This can return a handle.

anyway I fixed the texture because it wasnt loading things properly/the functionality I added was fundementally broken