cluster prepass DONE
points are being drawn to screen for testing. they show the bounding boxes for the clusters

When I identify active clusters, I'll use an index buffer to select which points to display and it will all be perfect.

the height and width values for AppWindow and Texture have been changed to size

alot of clean up is going to be necessary to make the code more legible

I've started using a different style in the class order. but I'll change all the different classes after I've finished the rendering so as  not to confuse the changes.