MSHHHHHHHHH thats my brain. my brain is moosh mushey mushé
Brrrrrrr

mesh size is managed by the vao now. it updates the size whenever a vbo is attached or detached. This means a buffer can resize and the vao wont notice. not ideal. so theres also a function to call an update on the size. its kinda hacky and ugly. not a fan. the alternative is to do it every draw call. That has to be worse. ughhh.

its not flawless it, like everything else needs tidying up.