Transform update 3001
I definitely havent changed Transforms enough.

Transform now uses tags instead of flags. I've seperated Position, Rotation, Scale, Local and World Transforms into seperate components.

This can save a calculation when a entity doesnt have all the transform attributes. But mostly I did it so that I could copy the Attributes of the Transforms into a GLbuffer with ease. I intend to do this with PointLights when i get to it.