Deep rock would love this one(there are alot of bugs) Deferred Clustered Lighting DONE(ish) New Syntax for mesh I put in namespace similar to Transform.
I deleted the old shaders and renamed 'clustered_shading' to 'deferred_shading'. it should really be called clustered_deferred_shading but oh well.

I've also renamed the shader Program to better reflect what they do.(and have a naming convention that doesnt just try and keep the variables the same length)

The framebuffer is stored in a sub class of RenderSystem called GeometryBuffer. Its ugly.
Render System needs cleaning up and that will go first.

changes needed to be made to texture to allow for un normalized data, previously it only supported RGBA and RGBAF(float). so I've made alot of changes to Texture. it now initiates on the default constructor and you can use setData() to add values.

apparently if you dont set the wrap and filter of a texture it breaks. This was a bug I spent ages on. very frustating.