i like the cut of your jib sir AHHHHHHHHHHHHHHHHHHH why why is that the fix thats so infuriating ughhhhhh
theres a really annoying stupididty with the material VBO. its strange. it only works if i input the materialIndex as a GL_FLOAT??? like what??? does it think it knows that a float is stored there so doesnt change it maybe???

theres also a wierd thing to do with the bindless textures. they arent using the correct sampling method. I thought it would use the one associated with the texture but it isnt.

anyways I changed a bunch Ill go in order.
in the geometry_pass.frag I've manually wrapped the texture coord to fix the problem of bindless textures forgetting how.

I've changed the vertex buffer materialID binding slot to 4 skipping over the bone ID and weight. this changes geometry_pass.vert and []

material is a reference container to make sure the stuff isnt deleted. it stores an array of structs called Material::Data which contains the Texture obj. it also stores an array of Material::Index. this stores a reference to the unique material ID in the buffer. this is looked up using ResourceManager<>.

in Resource.h GL<T> has been renamed to Shared<T> and FileManager has been renamed to ResourceManager<>. This is because I've added functionality where you can pass a custom handle type. it defaults to string for file usage but it's used by the Material::Index with size_t ie a hash value to look up if the index already exists.

I isolated the Framebuffer functionality and wrapped in a class.

Texture now stores its format. i did this so i could automatically resize the framebuffer attachments.

Load system now loads the textures. this process is kinda involved... first it collects all the material datas it hashes each material::Data then looks up to see if it already exists. if it does it makes a reference copy of the Index otherwise a new Index is created. in MaterialSystem there are 2 systems that process and assign the values into the OpenGL material buffer. this could be done better.

In Mesh System I've done the awkward fix for with GL_FLOAT on the material attach function.

I'm trying to use more registry::ctx which stands for context (I think). because it allows me to get the system member fields in events.

Ive loaded the sponza model and I'm using it for minor testing. I have nothing to really compare them to and I'm not sure I believe them anyway.

supposedly the scene is being rendered in less than a 1ms but if thats true its incredibly impressive. Alot of this might be because of the fairly restrictive limitations I gave to the light cluster size. thats whats giving the tiled artefacts you can see when you add a bajillion lights.

Oh I also removed the terrible set function from scene.

I'm nearing the point in this project where I want to do drastic changes. I've been reading up on vulkan and it looks alot more ordered and more pleasant(if dangerous) to use.